requirements.txt
asgiref==3.2.10
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
certifi==2020.4.5.2
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.7
django-cors-headers==3.4.0
django-eventstream==3.1.0
django-grip==3.0.0
djangorestframework==3.11.0
gripcontrol==4.0.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
psycopg2-binary==2.8.5
pubcontrol==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2020.1
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
Werkzeug==0.16.1
zope.interface==5.1.0
